Updated validation script to report on unused ScreenOS Service entries.
Updated the README file to reflect new functionality.
Added a new class, validation function, regexes, and configuration file
function to work on service entries.
Updated the reporting function to report on unused service entries and
to only print the report header when there are results to print.
Updated the policy regexes to the same as those used in the policy view script.
Minor updates here and there to function properly with the new policy
regexes.
Added statements to flush stdout to indicate the script was doing work.